requirements.txt
feedparser
pyyaml
requests
feedgen
python-slugify
edge-tts
